[FIX] website_[sale,delivery]: update freeshipping on delivery change
Since those two modules do not have a bridge, there was no support for changing the freeshipping line
when changing the delivery carrier

Now it has, at the price of ugly code

OPW 1871526
