+++ addons/website_sale_delivery/i18n/website_sale_delivery.pot
@@ -94,7 +94,10 @@ msgid "FedEx Delivery Methods"
 msgstr ""
 
 #. module: website_sale_delivery
+#. openerp-web
+#: code:addons/website_sale_delivery/static/src/js/website_sale_delivery.js:27
 #: model:ir.ui.view,arch_db:website_sale_delivery.payment_delivery
+#, python-format
 msgid "Free"
 msgstr ""
